Changed uploader logic and schema
Simplified the schema and populate the feeds table before starting the
loop.

This way we don't have duplicated informations and the schema is more
flexible because it can accomodate more feeds without the need of extending
it.

Also, the ->upload_feed method is more encapsulated and doesn't need to write
into the job table.